chore(structure): move into pkg directory
